+++ src/Engine/IndexBuffer.hpp
@@ -10,9 +10,9 @@ public:
     virtual ~IndexBuffer() = default;
     virtual void Bind() const = 0;
     virtual void Unbind() const = 0;
-    virtual uint32_t GetCount() const = 0;
+    virtual size_t GetCount() const = 0;
     
-    static Ref<IndexBuffer> Create(uint32_t* indices, uint32_t count);
+    static Ref<IndexBuffer> Create(uint32_t* indices, size_t count);
 };
 
 }
