+++ src/Engine/OpenGL3/OpenGL3VertexBuffer.cpp
@@ -6,14 +6,14 @@
 namespace OGLSample
 {
 
-OpenGL3VertexBuffer::OpenGL3VertexBuffer(uint32_t size)
+OpenGL3VertexBuffer::OpenGL3VertexBuffer(size_t size)
 : m_Size(size)
 {
     glGenBuffers(1, &m_Id);
     glBindBuffer(GL_ARRAY_BUFFER, m_Id);
     glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
 }
-OpenGL3VertexBuffer::OpenGL3VertexBuffer(float* vertices, uint32_t size)
+OpenGL3VertexBuffer::OpenGL3VertexBuffer(float* vertices, size_t size)
 : m_Size(size)
 {
     glGenBuffers(1, &m_Id);
@@ -35,13 +35,13 @@ void OpenGL3VertexBuffer::Unbind() const
     glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
 
-void OpenGL3VertexBuffer::SetData(const void* data, uint32_t size)
+void OpenGL3VertexBuffer::SetData(const void* data, size_t size)
 {
     glBindBuffer(GL_ARRAY_BUFFER, m_Id);
     glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
 }
 
-uint32_t OpenGL3VertexBuffer::GetCount()
+size_t OpenGL3VertexBuffer::GetCount()
 {
     return m_Size / m_Layout.GetStride();
 }
