+++ src/Engine/VertexBuffer.hpp
@@ -15,15 +15,15 @@ public:
     virtual void Bind() const = 0;
     virtual void Unbind() const = 0;
     
-    virtual void SetData(const void* data, uint32_t size) = 0;
+    virtual void SetData(const void* data, size_t size) = 0;
     
-    virtual uint32_t GetCount() = 0;
+    virtual size_t GetCount() = 0;
     
     virtual VertexBufferLayout& GetLayout() = 0;
     virtual void SetLayout(const VertexBufferLayout& layout) = 0;
     
-    static Ref<VertexBuffer> Create(uint32_t size);
-    static Ref<VertexBuffer> Create(float* vertices, uint32_t size);
+    static Ref<VertexBuffer> Create(size_t size);
+    static Ref<VertexBuffer> Create(float* vertices, size_t size);
 };
 
 }
